requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiohttp-cors==0.7.0
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blessed==1.19.1
blis==0.7.9
blosc2==2.0.0
bootstrapped==0.0.2
bs4==0.0.1
cachetools==5.2.1
catalogue==2.0.8
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorful==0.5.5
comm==0.1.2
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
Cython==0.29.33
dataclasses-json==0.5.7
debugpy==1.6.5
decorator==5.1.1
deepdiff==6.2.3
defusedxml==0.7.1
distlib==0.3.6
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
experiment-impact-tracker==0.1.8
fastjsonschema==2.16.2
filelock==3.9.0
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.3.0
google-api-core==2.11.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-cloud-storage==2.8.0
googleapis-common-protos==1.58.0
gpustat==1.0.0
grpcio==1.51.1
h5py==3.8.0
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.20.2
IPython>=8.10
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.6.2
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kaggle==1.5.12
kiwisolver==1.4.4
langcodes==3.3.0
ludwig==0.6.4
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-dataclass==8.5.5
marshmallow-enum==1.5.1
marshmallow-jsonschema==0.13.0
matplotlib==3.6.3
matplotlib-inline==0.1.6
mypy-boto3-sqs==1.26
mistune==2.0.4
msgpack==1.0.4
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.8
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.24.1
nvidia-ml-py==11.495.46
oauthlib==3.2.2
opencensus==0.11.1
opencensus-context==0.1.3
openpyxl==3.0.10
ordered-set==4.1.0
orjson==3.8.5
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
preshed==3.0.8
progiter==1.1.0
prometheus-client==0.13.1
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptitprince==0.2.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py-spy==0.3.14
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4
pyDeprecate==0.3.2
Pygments==2.14.0
PyHamcrest==2.0.4
PyLaTeX==1.4.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
python-slugify==8.0.0
pytz==2022.7.1
PyYAML==6.0
pyyaml==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
ratelim==0.1.6
ray==2.3.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.11.0
Send2Trash==1.8.0
shapely==2.0.0
six==1.16.0
slack-bolt==1.16.1
slack-sdk==3.19.5
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.5
stack-data==0.6.2
tables==3.8.0
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
text-unidecode==1.3
thinc==8.1.7
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.12.1
torch==1.13.1
torchaudio==0.13.1
torchinfo==1.7.1
torchmetrics==0.8.2
torchtext==0.14.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
transformers==4.21.3
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.4.0
ujson==5.7.0
uri-template==1.2.0
urllib3==1.26.14
virtualenv==20.17.1
wasabi==1.1.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug ~> 2.2.3
# Suggested by https://github.com/Determinant-AI/determinant/security/dependabot/2
widgetsnbextension>=2.2.3
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.2
zipp==3.11.0
fastapi==0.95.0
fakeredis==2.10.3
redis==4.5.4
